+++ src/main/java/com/picimako/mockitools/PointersUtil.java
@@ -4,24 +4,24 @@ package com.picimako.mockitools;
 
 import static java.util.stream.Collectors.toList;
 
-import java.util.List;
-
-import com.intellij.psi.PsiMethodCallExpression;
+import com.intellij.psi.PsiElement;
 import com.intellij.psi.SmartPointerManager;
 import com.intellij.psi.SmartPsiElementPointer;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 /**
  * Utility for {@link SmartPsiElementPointer}s.
  */
 public final class PointersUtil {
 
-    public static List<SmartPsiElementPointer<PsiMethodCallExpression>> toPointers(List<PsiMethodCallExpression> calls) {
+    public static <T extends PsiElement> List<SmartPsiElementPointer<T>> toPointers(List<T> calls) {
         return calls.stream().map(PointersUtil::toPointer).collect(toList());
     }
 
     @NotNull
-    public static SmartPsiElementPointer<PsiMethodCallExpression> toPointer(PsiMethodCallExpression element) {
+    public static <T extends PsiElement> SmartPsiElementPointer<T> toPointer(T element) {
         return SmartPointerManager.getInstance(element.getProject()).createSmartPsiElementPointer(element, element.getContainingFile());
     }