+++ src/test/java/com/picimako/mockitools/inspection/ThrowsCheckedExceptionStubbingInspectionTest.java
@@ -17,4 +17,74 @@ public class ThrowsCheckedExceptionStubbingInspectionTest extends MockitoolsInsp
     public void testThrowsCheckedExceptionStubbing() {
         doJavaTest();
     }
+
+    //Quick fixes
+
+    public void testAddsExceptionToEmptyThrowsClause() {
+        doQuickFixTest("Add exception to throws clause", "QuickFix.java",
+            "import java.io.IOException;\n" +
+                "import java.lang.NoSuchMethodException;\n" +
+                "import org.mockito.Mockito;\n" +
+                "class QuickFix {\n" +
+                "    void testMethod() {\n" +
+                "        MockObject mock = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.when(mock.doSomething()).thenThrow(IOExcep<caret>tion.class);\n" +
+                "    }\n" +
+                "\n" +
+                "    private static class MockObject {\n" +
+                "        public int doSomething() {\n" +
+                "            return 0;\n" +
+                "        }\n" +
+                "    }\n" +
+                "}",
+            "import java.io.IOException;\n" +
+                "import java.lang.NoSuchMethodException;\n" +
+                "import org.mockito.Mockito;\n" +
+                "class QuickFix {\n" +
+                "    void testMethod() {\n" +
+                "        MockObject mock = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.when(mock.doSomething()).thenThrow(IOException.class);\n" +
+                "    }\n" +
+                "\n" +
+                "    private static class MockObject {\n" +
+                "        public int doSomething() throws IOException {\n" +
+                "            return 0;\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testAddsExceptionToNonEmptyThrowsClause() {
+        doQuickFixTest("Add exception to throws clause", "QuickFix.java",
+            "import java.io.IOException;\n" +
+                "import java.lang.NoSuchMethodException;\n" +
+                "import org.mockito.Mockito;\n" +
+                "class QuickFix {\n" +
+                "    void testMethod() {\n" +
+                "        MockObject mock = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.when(mock.doSomething()).thenThrow(IOExcep<caret>tion.class);\n" +
+                "    }\n" +
+                "\n" +
+                "    private static class MockObject {\n" +
+                "        public int doSomething() throws NoSuchMethodException {\n" +
+                "            return 0;\n" +
+                "        }\n" +
+                "    }\n" +
+                "}",
+            "import java.io.IOException;\n" +
+                "import java.lang.NoSuchMethodException;\n" +
+                "import org.mockito.Mockito;\n" +
+                "class QuickFix {\n" +
+                "    void testMethod() {\n" +
+                "        MockObject mock = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.when(mock.doSomething()).thenThrow(IOException.class);\n" +
+                "    }\n" +
+                "\n" +
+                "    private static class MockObject {\n" +
+                "        public int doSomething() throws NoSuchMethodException, IOException {\n" +
+                "            return 0;\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
 }