+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ## [0.5.0]
 ### Added
-- Strictness configuration is now supported by the conversion between `@Mock` and `Mockito.mock(Class, MockSettings)` in both directions.
+- [#24](https://github.com/picimako/mockitools/issues/24): Strictness configuration is now supported by the conversion between `@Mock` and `Mockito.mock(Class, MockSettings)` in both directions.
+- [#23](https://github.com/picimako/mockitools/issues/23): Added a quick fix, so that checked exceptions can be added to the stubbed method's `throws` clause
+when they are specified in the stubbing, but not in the mentioned `throws` clause.
 
 ### Changed
 - Removed support for IJ2021.1. From now on 2021.2 is the earliest version supported.